updated difficulties to include ddrcommunity.com tiering guides
The intent of this change is for training/practice/getting better, not tournament play. All the singles difficulties between 13-18 are no longer whole numbers. For example, it would allow a user to generate a random set of songs 14.6-15.4 when they are just starting to pass 15s or are grinding out PFCs on hard 14s/easy 15s.

In order to use this version you need to be careful to remember that difficulties are now float values and if you want to look at the entire range of a number as an upper bound you will need to use the formatting "#.99", otherwise you will only return songs that are rated #.00.